Switched plugin structure to heedy-template-plugin
This makes life easier, since it unifies plugin structure, and makes debugging cleaner, sicne the plugin template includes useful features.

+++ Makefile
@@ -1,39 +1,46 @@
-VERSION:=$(shell cat VERSION)-git.$(shell git rev-list --count HEAD)
+VERSION:=$(shell cat VERSION)
+PLUGIN_NAME:=notebook
 
-.PHONY: clean phony
+HEEDY:=$(shell test -f ./heedy && echo "./")heedy
 
-all: build
+.PHONY: clean clear watch debug release run
 
-#Empty rule for forcing rebuilds
-phony:
 
-build: frontend
-	# Remove original mjs files in production builds
-	find ./dist/notebook/public/static -name "*.gz" -exec sh -c 'rm "$${0%.gz}"' {} ';'
+release: node_modules clear dist/notebook
+	npm run build
+	# Remove original files of pre-compressed content
+	if test -d ./dist/notebook/public/static; then find ./dist/notebook/public/static -name "*.gz" -exec sh -c 'rm "$${0%.gz}"' {} ';';fi
+	# Zip the plugin
 	cd dist;zip -r heedy-notebook-plugin-${VERSION}.zip ./notebook
 
-dist/notebook: node_modules
-	mkdir -p dist/notebook
-	cp LICENSE ./dist/notebook
-	cp VERSION ./dist/notebook
-	cp requirements.txt ./dist/notebook
-	cp -r ipynb ./dist/notebook
-	npm run build
-
 node_modules:
+	if test -d frontend; then (cd frontend && npm i);fi
 	npm i
 
-frontend/node_modules:
-	cd frontend; npm i
+dist/notebook:
+	mkdir -p dist/notebook
 
-frontend: phony dist/notebook frontend/node_modules
-	cd frontend; npm run build
+testdb:
+	$(HEEDY) create testdb --noserver --username=test --password=test
+	mkdir testdb/plugins
+	cd testdb/plugins; ln -s ../../dist/notebook notebook
 
-	
-debug: phony frontend/node_modules dist/notebook
+debug: node_modules dist/notebook testdb
 	npm run debug
 
-clean:
-	rm -rf ./dist
-	rm -rf ./node_modules
-	rm -rf ./frontend/node_modules
+watch: node_modules dist/notebook testdb
+	npm run watch
+
+run: testdb
+	$(HEEDY) run testdb --verbose
+
+clear:
+	rm -rf dist
+
+clean: clear
+	rm -rf testdb
+	if test -d node_modules; then rm -rf node_modules; fi
+	if test -d frontend/node_modules; then rm -rf frontend/node_modules; fi
+
+rename: clean
+	find ./ -type f -exec sed -i -e 's/notebook/$(PLUGIN_NAME)/g' {} \;
